refactor: re-organize code files for JS
