+++ index.html
@@ -6,6 +6,6 @@
 </head>
 <body style="margin:0;padding:0;height:100%;background:#1b2632;overflow:hidden">
     <div id="game"></div>
-    <script src="./phaser.min.js"></script>
-    <script type="module" src="./game.js"></script>
+    <script src="./script-phaser.js"></script>
+    <script type="module" src="./script-game.js"></script>
 </body>
